Pool common fixtures to package conftest
This removes the need to import fixtures as unused variables in the
other test packages. The autoflake hook removed those unused imports,
and with this, we don't need to import them in the first place.
